Fix for fpga
