+++ dlk/python/dlk/templates/CMakeLists.txt
@@ -34,6 +34,7 @@ list(APPEND SRC_LIB_ALL
     src/func/conv2d.cpp
     src/func/impl/apply_thresholds.cpp
     src/func/impl/quantized_conv2d_dim2col.cpp
+    src/func/impl/ohwi_to_hwoi.cpp
     src/func/max_pool.cpp
     src/func/pad.cpp
     src/func/matmul.cpp
