+++ dlk/python/dlk/templates/src/func/impl/generic/quantized_conv2d_kn2row.cpp
@@ -24,37 +24,6 @@ limitations under the License.
 #include "pack_input_to_qwords.h"
 #include "time_measurement.h"
 
-namespace dlk {
-
-namespace impl {
-
-// kernel format converter
-// ohwi : oc kh kw ic, hwoi: kh kw oc ic
-void quantized_ohwi_to_hwoi(const kernel_t& ohwi, const kn2row_kernel_t& hwoi, const binary_convolution_parameters& p) {
-  Measurement::Start("quantized_ohwi_to_hwoi");
-
-  int ic = p.normal_conv_params.kernel_depth / 32;
-  int oc = p.normal_conv_params.output_channels;
-  int kh = p.normal_conv_params.kernel_height;
-  int kw = p.normal_conv_params.kernel_width;
-
-  for (unsigned int i = 0; i < kh; ++i) {
-    for (unsigned int j = 0; j < kw; ++j) {
-      for (unsigned int k = 0; k < oc; ++k) {
-        for (unsigned int l = 0; l < ic; ++l) {
-          hwoi(i, j, k, l) = ohwi(k, i, j, l);
-        }
-      }
-    }
-  }
-
-  Measurement::Stop();
-}
-
-} // namespace impl
-
-} // namespace dlk
-
 namespace {
 
 void ApplyThresholds(
